lambda-s3-trigger/ingestion-worker/requirements.txt
boto3==1.34.0
pandas==2.1.4
psycopg2-binary==2.9.9
torch==2.1.2
transformers==4.37.0
numpy>=1.24.0,<2.0.0
sentence-transformers==2.2.2
